requirements.txt
streamlit
pandas
plotly
matplotlib
seaborn
pillow